+++ LoopKitUI/Views/Settings Editors/TherapySettingsViewModel.swift
@@ -7,8 +7,9 @@
 //
 
 import Combine
-import LoopKit
 import HealthKit
+import LocalAuthentication
+import LoopKit
 import SwiftUI
 
 public class TherapySettingsViewModel: ObservableObject {
@@ -26,6 +27,7 @@ public class TherapySettingsViewModel: ObservableObject {
     let sensitivityOverridesEnabled: Bool
     let prescription: Prescription?
     let appName: String
+    let authenticationChallengeDescription: String
 
     lazy private var cancellables = Set<AnyCancellable>()
 
@@ -37,6 +39,7 @@ public class TherapySettingsViewModel: ObservableObject {
                 syncPumpSchedule: PumpManager.SyncSchedule? = nil,
                 sensitivityOverridesEnabled: Bool = false,
                 prescription: Prescription? = nil,
+                authenticationChallengeDescription: String = "Authenticate to change setting",
                 didSave: SaveCompletion? = nil) {
         self.mode = mode
         self.therapySettings = therapySettings
@@ -47,6 +50,7 @@ public class TherapySettingsViewModel: ObservableObject {
         self.prescription = prescription
         self.supportedInsulinModelSettings = supportedInsulinModelSettings
         self.appName = appName
+        self.authenticationChallengeDescription = authenticationChallengeDescription
         self.didSave = didSave
     }
     
@@ -69,43 +73,64 @@ public class TherapySettingsViewModel: ObservableObject {
     
     public func saveCorrectionRange(range: GlucoseRangeSchedule) {
         therapySettings.glucoseTargetRangeSchedule = range
-        didSave?(TherapySetting.glucoseTargetRange, therapySettings)
+        beginSaving(TherapySetting.glucoseTargetRange)
     }
     
     public func saveCorrectionRangeOverrides(overrides: CorrectionRangeOverrides, unit: HKUnit) {
         therapySettings.preMealTargetRange = overrides.preMeal?.doubleRange(for: unit)
         therapySettings.workoutTargetRange = overrides.workout?.doubleRange(for: unit)
-        didSave?(TherapySetting.correctionRangeOverrides, therapySettings)
+        beginSaving(TherapySetting.correctionRangeOverrides)
     }
     
     public func saveSuspendThreshold(value: GlucoseThreshold) {
         therapySettings.suspendThreshold = value
-        didSave?(TherapySetting.suspendThreshold, therapySettings)
+        beginSaving(TherapySetting.suspendThreshold)
     }
     
     public func saveBasalRates(basalRates: BasalRateSchedule) {
         therapySettings.basalRateSchedule = basalRates
-        didSave?(TherapySetting.basalRate, therapySettings)
+        beginSaving(TherapySetting.basalRate)
     }
     
     public func saveDeliveryLimits(limits: DeliveryLimits) {
         therapySettings.maximumBasalRatePerHour = limits.maximumBasalRate?.doubleValue(for: .internationalUnitsPerHour)
         therapySettings.maximumBolus = limits.maximumBolus?.doubleValue(for: .internationalUnit())
-        didSave?(TherapySetting.deliveryLimits, therapySettings)
+        beginSaving(TherapySetting.deliveryLimits)
     }
     
     public func saveInsulinModel(insulinModelSettings: InsulinModelSettings) {
         therapySettings.insulinModelSettings = insulinModelSettings
-        didSave?(TherapySetting.insulinModel, therapySettings)
+        beginSaving(TherapySetting.insulinModel)
     }
     
     public func saveCarbRatioSchedule(carbRatioSchedule: CarbRatioSchedule) {
         therapySettings.carbRatioSchedule = carbRatioSchedule
-        didSave?(TherapySetting.carbRatio, therapySettings)
+        beginSaving(TherapySetting.carbRatio)
     }
     
     public func saveInsulinSensitivitySchedule(insulinSensitivitySchedule: InsulinSensitivitySchedule) {
         therapySettings.insulinSensitivitySchedule = insulinSensitivitySchedule
-        didSave?(TherapySetting.insulinSensitivity, therapySettings)
+        beginSaving(TherapySetting.insulinSensitivity)
+    }
+    
+    private func beginSaving(_ setting: TherapySetting) {
+        let context = LAContext()
+        if context.canEvaluatePolicy(.deviceOwnerAuthentication, error: nil) {
+            context.evaluatePolicy(.deviceOwnerAuthentication,
+                                   localizedReason: authenticationChallengeDescription,
+                                   reply: { (success, error) in
+                                    if success {
+                                        DispatchQueue.main.async {
+                                            self.continueSaving(setting)
+                                        }
+                                    }
+            })
+        } else {
+            self.continueSaving(setting)
+        }
+    }
+    
+    private func continueSaving(_ setting: TherapySetting) {
+        didSave?(setting, therapySettings)
     }
 }
